chore: move async query functions to its own service
