+++ packages/backend/src/ee/index.ts
@@ -3,6 +3,7 @@ import express, { Express } from 'express';
 import { AppArguments } from '../App';
 import { lightdashConfig } from '../config/lightdashConfig';
 import Logger from '../logging/logger';
+import { AsyncQueryService } from '../services/AsyncQueryService/AsyncQueryService';
 import { ProjectService } from '../services/ProjectService/ProjectService';
 import { EncryptionUtil } from '../utils/EncryptionUtil/EncryptionUtil';
 import LicenseClient from './clients/License/LicenseClient';
@@ -138,14 +139,8 @@ export async function getEnterpriseAppArguments(): Promise<EnterpriseAppArgument
                     projectService: repository.getProjectService(),
                     lightdashConfig: context.lightdashConfig,
                 }),
-            cacheService: ({ models, context, clients }) =>
-                new CommercialCacheService({
-                    resultsCacheModel: models.getResultsCacheModel(),
-                    lightdashConfig: context.lightdashConfig,
-                    storageClient: clients.getResultsCacheStorageClient(),
-                }),
-            projectService: ({ models, context, clients, utils, repository }) =>
-                new ProjectService<IResultsCacheStorageClient>({
+            projectService: ({ models, context, clients, utils }) =>
+                new ProjectService({
                     lightdashConfig: context.lightdashConfig,
                     analytics: context.lightdashAnalytics,
                     projectModel: models.getProjectModel(),
@@ -172,9 +167,51 @@ export async function getEnterpriseAppArguments(): Promise<EnterpriseAppArgument
                     catalogModel: models.getCatalogModel(),
                     contentModel: models.getContentModel(),
                     encryptionUtil: utils.getEncryptionUtil(),
-                    queryHistoryModel: models.getQueryHistoryModel(),
                     userModel: models.getUserModel(),
+                }),
+            asyncQueryService: ({
+                models,
+                context,
+                clients,
+                utils,
+                repository,
+            }) =>
+                new AsyncQueryService<IResultsCacheStorageClient>({
+                    lightdashConfig: context.lightdashConfig,
+                    analytics: context.lightdashAnalytics,
+                    projectModel: models.getProjectModel(),
+                    onboardingModel: models.getOnboardingModel(),
+                    savedChartModel: models.getSavedChartModel(),
+                    jobModel: models.getJobModel(),
+                    emailClient: clients.getEmailClient(),
+                    spaceModel: models.getSpaceModel(),
+                    sshKeyPairModel: models.getSshKeyPairModel(),
+                    userAttributesModel: models.getUserAttributesModel(),
+                    s3CacheClient: clients.getS3CacheClient(),
+                    analyticsModel: models.getAnalyticsModel(),
+                    dashboardModel: models.getDashboardModel(),
+                    userWarehouseCredentialsModel:
+                        models.getUserWarehouseCredentialsModel(),
+                    warehouseAvailableTablesModel:
+                        models.getWarehouseAvailableTablesModel(),
+                    emailModel: models.getEmailModel(),
+                    schedulerClient: clients.getSchedulerClient(),
+                    downloadFileModel: models.getDownloadFileModel(),
+                    s3Client: clients.getS3Client(),
+                    groupsModel: models.getGroupsModel(),
+                    tagsModel: models.getTagsModel(),
+                    catalogModel: models.getCatalogModel(),
+                    contentModel: models.getContentModel(),
+                    encryptionUtil: utils.getEncryptionUtil(),
+                    userModel: models.getUserModel(),
+                    queryHistoryModel: models.getQueryHistoryModel(),
                     cacheService: repository.getCacheService(),
+                }),
+            cacheService: ({ models, context, clients }) =>
+                new CommercialCacheService({
+                    resultsCacheModel: models.getResultsCacheModel(),
+                    lightdashConfig: context.lightdashConfig,
+                    storageClient: clients.getResultsCacheStorageClient(),
                 }),
         },
         modelProviders: {
