+++ packages/backend/src/services/ProjectService/ProjectService.ts
@@ -7,8 +7,6 @@ import {
     AnyType,
     ApiChartAndResults,
     type ApiCreateProjectResults,
-    type ApiExecuteAsyncQueryResults,
-    type ApiGetAsyncQueryResults,
     ApiQueryResults,
     ApiSqlQueryResults,
     assertUnreachable,
@@ -21,7 +19,6 @@ import {
     ContentType,
     convertCustomMetricToDbt,
     convertExplores,
-    convertFieldRefToFieldId,
     countCustomDimensionsInMetricQuery,
     countTotalFilterRules,
     type CreateDatabricksCredentials,
@@ -30,10 +27,8 @@ import {
     CreateProject,
     CreateProjectMember,
     CreateSnowflakeCredentials,
-    createVirtualView as createVirtualViewObject,
     CreateVirtualViewPayload,
     CreateWarehouseCredentials,
-    type CustomDimension,
     CustomFormatType,
     DashboardAvailableFilters,
     DashboardBasicDetails,
@@ -46,15 +41,9 @@ import {
     DbtProjectType,
     DbtRawModelNode,
     deepEqual,
-    DEFAULT_RESULTS_PAGE_SIZE,
     DefaultSupportedDbtVersion,
     DimensionType,
     DownloadFileType,
-    type ExecuteAsyncDashboardChartRequestParams,
-    type ExecuteAsyncMetricQueryRequestParams,
-    type ExecuteAsyncQueryRequestParams,
-    type ExecuteAsyncSavedChartRequestParams,
-    type ExecuteAsyncUnderlyingDataRequestParams,
     Explore,
     ExploreError,
     ExploreType,
@@ -65,7 +54,6 @@ import {
     findReplaceableCustomMetrics,
     ForbiddenError,
     formatRawRows,
-    formatRow,
     formatRows,
     getAggregatedField,
     getDashboardFilterRulesForTables,
@@ -79,19 +67,14 @@ import {
     getMetrics,
     getSubtotalKey,
     getTimezoneLabel,
-    GroupByColumn,
     hasIntersection,
     IntrinsicUserAttributes,
-    isCustomBinDimension,
-    isCustomDimension,
     isCustomSqlDimension,
     isDateItem,
     isDimension,
     isExploreError,
-    isField,
     isFilterableDimension,
     isFilterRule,
-    isMetric,
     isNotNull,
     isUserWithOrg,
     ItemsMap,
@@ -111,9 +94,7 @@ import {
     NotFoundError,
     ParameterError,
     PivotChartData,
-    PivotIndexColum,
     PivotValuesColumn,
-    prefixPivotConfigurationReferences,
     Project,
     ProjectCatalog,
     ProjectGroupAccess,
@@ -121,7 +102,6 @@ import {
     ProjectMemberRole,
     ProjectType,
     QueryExecutionContext,
-    QueryHistoryStatus,
     ReplaceableCustomFields,
     ReplaceCustomFields,
     ReplaceCustomFieldsPayload,
@@ -134,7 +114,6 @@ import {
     type SemanticLayerConnectionUpdate,
     SessionUser,
     snakeCaseName,
-    SortBy,
     SortByDirection,
     SortField,
     SpaceQuery,
@@ -152,7 +131,6 @@ import {
     UpdateVirtualViewPayload,
     UserAttributeValueMap,
     UserWarehouseCredentials,
-    ValuesColumn,
     VizColumn,
     WarehouseClient,
     WarehouseCredentials,
@@ -193,7 +171,6 @@ import { GroupsModel } from '../../models/GroupsModel';
 import { JobModel } from '../../models/JobModel/JobModel';
 import { OnboardingModel } from '../../models/OnboardingModel/OnboardingModel';
 import { ProjectModel } from '../../models/ProjectModel/ProjectModel';
-import { QueryHistoryModel } from '../../models/QueryHistoryModel';
 import { SavedChartModel } from '../../models/SavedChartModel';
 import { SpaceModel } from '../../models/SpaceModel';
 import { SshKeyPairModel } from '../../models/SshKeyPairModel';
@@ -215,12 +192,6 @@ import { ProjectAdapter } from '../../types';
 import { runWorkerThread, wrapSentryTransaction } from '../../utils';
 import { EncryptionUtil } from '../../utils/EncryptionUtil/EncryptionUtil';
 import { BaseService } from '../BaseService';
-import type { ICacheService } from '../CacheService/ICacheService';
-import {
-    CreateCacheResult,
-    MissCacheResult,
-    ResultsCacheStatus,
-} from '../CacheService/types';
 import {
     hasDirectAccessToSpace,
     hasViewAccessToSpace,
@@ -231,21 +202,8 @@ import {
     getFilteredExplore,
 } from '../UserAttributesService/UserAttributeUtils';
 import { ValidationService } from '../ValidationService/ValidationService';
-import {
-    getNextAndPreviousPage,
-    validatePagination,
-} from './resultsPagination';
-import {
-    type ExecuteAsyncDashboardChartQueryArgs,
-    type ExecuteAsyncMetricQueryArgs,
-    type ExecuteAsyncQueryReturn,
-    type ExecuteAsyncSavedChartQueryArgs,
-    ExecuteAsyncSqlQueryArgs,
-    type ExecuteAsyncUnderlyingDataQueryArgs,
-    type GetAsyncQueryResultsArgs,
-} from './types';
 
-type ProjectServiceArguments<ResultsCacheStorageClient = unknown> = {
+export type ProjectServiceArguments = {
     lightdashConfig: LightdashConfig;
     analytics: LightdashAnalytics;
     projectModel: ProjectModel;
@@ -269,15 +227,11 @@ type ProjectServiceArguments<ResultsCacheStorageClient = unknown> = {
     tagsModel: TagsModel;
     catalogModel: CatalogModel;
     contentModel: ContentModel;
-    queryHistoryModel: QueryHistoryModel;
     encryptionUtil: EncryptionUtil;
     userModel: UserModel;
-    cacheService?: ICacheService<ResultsCacheStorageClient>;
 };
 
-export class ProjectService<
-    ResultsCacheStorageClient = unknown,
-> extends BaseService {
+export class ProjectService extends BaseService {
     lightdashConfig: LightdashConfig;
 
     analytics: LightdashAnalytics;
@@ -326,13 +280,9 @@ export class ProjectService<
 
     contentModel: ContentModel;
 
-    queryHistoryModel: QueryHistoryModel;
-
     encryptionUtil: EncryptionUtil;
 
     userModel: UserModel;
-
-    cacheService?: ICacheService<ResultsCacheStorageClient>;
 
     constructor({
         lightdashConfig,
@@ -358,11 +308,9 @@ export class ProjectService<
         tagsModel,
         catalogModel,
         contentModel,
-        queryHistoryModel,
         encryptionUtil,
         userModel,
-        cacheService,
-    }: ProjectServiceArguments<ResultsCacheStorageClient>) {
+    }: ProjectServiceArguments) {
         super();
         this.lightdashConfig = lightdashConfig;
         this.analytics = analytics;
@@ -388,10 +336,8 @@ export class ProjectService<
         this.tagsModel = tagsModel;
         this.catalogModel = catalogModel;
         this.contentModel = contentModel;
-        this.queryHistoryModel = queryHistoryModel;
         this.encryptionUtil = encryptionUtil;
         this.userModel = userModel;
-        this.cacheService = cacheService;
     }
 
     static getMetricQueryExecutionProperties({
@@ -639,8 +585,7 @@ export class ProjectService<
     }
 
     // TODO: getWarehouseCredentials could be moved to a client WarehouseClientManager. However, this client shouldn't be using a model. Perhaps this information can be passed as a prop to the client so that other services can use the warehouse client credentials logic?
-
-    private async getWarehouseCredentials(
+    protected async getWarehouseCredentials(
         projectUuid: string,
         userUuid: string,
     ) {
@@ -1966,1408 +1911,6 @@ export class ProjectService<
             chartUuid: undefined,
             invalidateCache: true, // Do not cache results for explore queries
         });
-    }
-
-    async cancelAsyncQuery({
-        user,
-        projectUuid,
-        queryUuid,
-    }: {
-        user: SessionUser;
-        projectUuid: string;
-        queryUuid: string;
-    }): Promise<void> {
-        const { organizationUuid } = await this.projectModel.getSummary(
-            projectUuid,
-        );
-        if (
-            user.ability.cannot(
-                'view',
-                subject('Project', { organizationUuid, projectUuid }),
-            )
-        ) {
-            throw new ForbiddenError();
-        }
-
-        const queryHistory = await this.queryHistoryModel.get(
-            queryUuid,
-            projectUuid,
-            user.userUuid,
-        );
-
-        if (user.userUuid !== queryHistory.createdByUserUuid) {
-            throw new ForbiddenError(
-                'User is not allowed to cancel this query',
-            );
-        }
-
-        await this.queryHistoryModel.update(
-            queryUuid,
-            projectUuid,
-            user.userUuid,
-            {
-                status: QueryHistoryStatus.CANCELLED,
-            },
-        );
-    }
-
-    async getAsyncQueryResults({
-        user,
-        projectUuid,
-        queryUuid,
-        page = 1,
-        pageSize,
-    }: GetAsyncQueryResultsArgs): Promise<ApiGetAsyncQueryResults> {
-        if (!isUserWithOrg(user)) {
-            throw new ForbiddenError('User is not part of an organization');
-        }
-        const { organizationUuid } =
-            await this.projectModel.getWithSensitiveFields(projectUuid);
-
-        const queryHistory = await this.queryHistoryModel.get(
-            queryUuid,
-            projectUuid,
-            user.userUuid,
-        );
-
-        if (
-            user.ability.cannot(
-                'view',
-                subject('Project', { organizationUuid, projectUuid }),
-            )
-        ) {
-            throw new ForbiddenError();
-        }
-
-        if (user.userUuid !== queryHistory.createdByUserUuid) {
-            throw new ForbiddenError(
-                'User is not allowed to fetch results for this query',
-            );
-        }
-
-        const { metricQuery, context, status, totalRowCount, cacheKey } =
-            queryHistory;
-
-        const defaultedPageSize =
-            pageSize ??
-            queryHistory.defaultPageSize ??
-            DEFAULT_RESULTS_PAGE_SIZE;
-
-        validatePagination({
-            pageSize: defaultedPageSize,
-            page,
-            queryMaxLimit: this.lightdashConfig.query.maxPageSize,
-            totalRowCount,
-        });
-
-        const resultsCacheEnabled =
-            this.lightdashConfig.resultsCache?.resultsEnabled;
-
-        switch (status) {
-            case QueryHistoryStatus.CANCELLED:
-                return {
-                    status,
-                    queryUuid,
-                    metricQuery,
-                    fields: queryHistory.fields,
-                };
-            case QueryHistoryStatus.PENDING:
-                if (resultsCacheEnabled && cacheKey && this.cacheService) {
-                    const cache = await this.cacheService.findCache(
-                        cacheKey,
-                        projectUuid,
-                    );
-
-                    if (cache?.status === ResultsCacheStatus.READY) {
-                        // If the cache is ready, we update the query history status to READY and return the current state
-                        // This avoids a race condition where the query history was READY but the cache (triggered by another request) was still being written
-                        this.logger.debug(
-                            `Updating query history status to READY for query ${queryUuid} in project ${projectUuid}. Cache status: ${cache?.status}`,
-                        );
-                        void this.queryHistoryModel.update(
-                            queryUuid,
-                            projectUuid,
-                            user.userUuid,
-                            {
-                                status: QueryHistoryStatus.READY,
-                                total_row_count: cache.totalRowCount,
-                            },
-                        );
-                    }
-                }
-
-                return {
-                    status,
-                    queryUuid,
-                    metricQuery,
-                    fields: queryHistory.fields,
-                };
-            case QueryHistoryStatus.ERROR:
-                return {
-                    status,
-                    queryUuid,
-                    metricQuery,
-                    fields: queryHistory.fields,
-                    error: queryHistory.error,
-                };
-            case QueryHistoryStatus.READY:
-                break;
-            default:
-                return assertUnreachable(status, 'Unknown query status');
-        }
-
-        const formatter = (row: Record<string, unknown>) =>
-            formatRow(row, queryHistory.fields);
-
-        let returnObject: ApiGetAsyncQueryResults;
-        let roundedDurationMs: number;
-        if (resultsCacheEnabled && cacheKey && this.cacheService) {
-            // Need to save to a variable, otherwise TS doesn't carry over the fact that this.cacheService is not undefined through the measureTime call
-            const { cacheService } = this;
-            const {
-                result: { rows, totalRowCount: cacheTotalRowCount, expiresAt },
-                durationMs,
-            } = await measureTime(
-                () =>
-                    cacheService.getCachedResultsPage(
-                        cacheKey,
-                        projectUuid,
-                        page,
-                        defaultedPageSize,
-                        formatter,
-                    ),
-                'getCachedResultsPage',
-                this.logger,
-                context,
-            );
-
-            const pageCount = Math.ceil(cacheTotalRowCount / defaultedPageSize);
-
-            roundedDurationMs = Math.round(durationMs);
-
-            const { nextPage, previousPage } = getNextAndPreviousPage(
-                page,
-                pageCount,
-            );
-
-            this.analytics.track({
-                userId: user.userUuid,
-                event: 'results_cache.read',
-                properties: {
-                    queryId: queryHistory.queryUuid,
-                    projectId: projectUuid,
-                    cacheKey,
-                    page,
-                    requestedPageSize: defaultedPageSize,
-                    rowCount: rows.length,
-                    resultsPageExecutionMs: roundedDurationMs,
-                },
-            });
-
-            this.analytics.track({
-                userId: user.userUuid,
-                event: 'query_page.fetched',
-                properties: {
-                    queryId: queryHistory.queryUuid,
-                    projectId: projectUuid,
-                    warehouseType:
-                        queryHistory?.warehouseQueryMetadata?.type ?? null,
-                    page,
-                    columnsCount: Object.keys(queryHistory.fields).length,
-                    totalRowCount: cacheTotalRowCount,
-                    totalPageCount: pageCount,
-                    resultsPageSize: rows.length,
-                    resultsPageExecutionMs: roundedDurationMs,
-                    status,
-                    cacheMetadata: {
-                        cacheExpiresAt: expiresAt,
-                        cacheKey,
-                    },
-                },
-            });
-
-            returnObject = {
-                rows,
-                totalPageCount: pageCount,
-                totalResults: cacheTotalRowCount,
-                queryUuid: queryHistory.queryUuid,
-                fields: queryHistory.fields,
-                metricQuery,
-                pageSize: rows.length,
-                page,
-                nextPage,
-                previousPage,
-                initialQueryExecutionMs:
-                    queryHistory.warehouseExecutionTimeMs ?? roundedDurationMs,
-                resultsPageExecutionMs: roundedDurationMs,
-                status,
-            };
-        } else {
-            const explore = await this.getExplore(
-                user,
-                projectUuid,
-                metricQuery.exploreName,
-            );
-
-            const { warehouseClient, sshTunnel } =
-                await this._getWarehouseClient(
-                    projectUuid,
-                    await this.getWarehouseCredentials(
-                        projectUuid,
-                        user.userUuid,
-                    ),
-                    {
-                        snowflakeVirtualWarehouse: explore.warehouse,
-                        databricksCompute: explore.databricksCompute,
-                    },
-                );
-
-            const queryTags: RunQueryTags = {
-                organization_uuid: organizationUuid,
-                project_uuid: projectUuid,
-                user_uuid: user.userUuid,
-                explore_name: metricQuery.exploreName,
-                query_context: context,
-            };
-
-            const { result, durationMs } = await measureTime(
-                () =>
-                    warehouseClient
-                        .getAsyncQueryResults(
-                            {
-                                page,
-                                pageSize: defaultedPageSize,
-                                tags: queryTags,
-                                queryId: queryHistory.warehouseQueryId,
-                                sql: queryHistory.compiledSql,
-                                queryMetadata:
-                                    queryHistory.warehouseQueryMetadata,
-                            },
-                            formatter,
-                        )
-                        .catch((e) => ({ errorMessage: getErrorMessage(e) })),
-                'getAsyncQueryResults',
-                this.logger,
-                context,
-            );
-
-            if ('errorMessage' in result) {
-                this.analytics.track({
-                    userId: user.userUuid,
-                    event: 'query.error',
-                    properties: {
-                        queryId: queryHistory.queryUuid,
-                        projectId: projectUuid,
-                        warehouseType: warehouseClient.credentials.type,
-                    },
-                });
-                await this.queryHistoryModel.update(
-                    queryHistory.queryUuid,
-                    projectUuid,
-                    user.userUuid,
-                    {
-                        status: QueryHistoryStatus.ERROR,
-                        error: result.errorMessage,
-                    },
-                );
-
-                return {
-                    status: QueryHistoryStatus.ERROR,
-                    error: result.errorMessage,
-                    queryUuid: queryHistory.queryUuid,
-                    metricQuery,
-                    fields: queryHistory.fields,
-                };
-            }
-
-            roundedDurationMs = Math.round(durationMs);
-
-            this.analytics.track({
-                userId: user.userUuid,
-                event: 'query_page.fetched',
-                properties: {
-                    queryId: queryHistory.queryUuid,
-                    projectId: projectUuid,
-                    warehouseType: warehouseClient.credentials.type,
-                    page,
-                    columnsCount: Object.keys(result.fields).length,
-                    totalRowCount: result.totalRows,
-                    totalPageCount: result.pageCount,
-                    resultsPageSize: result.rows.length,
-                    resultsPageExecutionMs: roundedDurationMs,
-                    status,
-                    cacheMetadata: null,
-                },
-            });
-
-            const { nextPage, previousPage } = getNextAndPreviousPage(
-                page,
-                result.pageCount,
-            );
-
-            returnObject = {
-                rows: result.rows,
-                totalPageCount: result.pageCount,
-                totalResults: result.totalRows,
-                queryUuid: queryHistory.queryUuid,
-                fields: queryHistory.fields,
-                metricQuery,
-                pageSize: result.rows.length,
-                page,
-                nextPage,
-                previousPage,
-                initialQueryExecutionMs:
-                    queryHistory.warehouseExecutionTimeMs ?? roundedDurationMs,
-                resultsPageExecutionMs: roundedDurationMs,
-                status,
-            };
-
-            await sshTunnel.disconnect();
-        }
-
-        /**
-         * Update the query history with non null values
-         * defaultPageSize is null when user never fetched the results - we don't send pagination params to the query execution endpoint
-         * warehouseExecutionTimeMs is null when warehouse doesn't support async queries - query is only executed when user fetches results
-         * totalRowCount is null when warehouse doesn't support async queries - query is only executed when user fetches results
-         */
-        if (
-            queryHistory.defaultPageSize === null ||
-            queryHistory.warehouseExecutionTimeMs === null ||
-            queryHistory.totalRowCount === null
-        ) {
-            await this.queryHistoryModel.update(
-                queryHistory.queryUuid,
-                projectUuid,
-                user.userUuid,
-                {
-                    ...(queryHistory.defaultPageSize === null
-                        ? { default_page_size: defaultedPageSize }
-                        : {}),
-                    ...(queryHistory.warehouseExecutionTimeMs === null
-                        ? {
-                              warehouse_execution_time_ms: roundedDurationMs,
-                          }
-                        : {}),
-                    ...(queryHistory.totalRowCount === null
-                        ? { total_row_count: returnObject.totalResults }
-                        : {}),
-                },
-            );
-        }
-
-        return returnObject;
-    }
-
-    /**
-     * Runs the query the warehouse and updates the query history and cache (if cache is enabled and cache is not hit) when complete
-     */
-    private async runAsyncWarehouseQuery({
-        user,
-        projectUuid,
-        query,
-        fieldsMap,
-        queryTags,
-        warehouseClient,
-        sshTunnel,
-        queryHistoryUuid,
-        resultsCache,
-    }: {
-        user: SessionUser;
-        projectUuid: string;
-        queryTags: RunQueryTags;
-        query: string;
-        fieldsMap: ItemsMap;
-        queryHistoryUuid: string;
-        resultsCache?: MissCacheResult;
-        warehouseClient: WarehouseClient;
-        sshTunnel: SshTunnel<CreateWarehouseCredentials>;
-    }) {
-        try {
-            const { queryId, queryMetadata, totalRows, durationMs } =
-                await warehouseClient.executeAsyncQuery(
-                    {
-                        sql: query,
-                        tags: queryTags,
-                    },
-                    resultsCache?.write,
-                );
-
-            this.analytics.track({
-                userId: user.userUuid,
-                event: 'query.ready',
-                properties: {
-                    queryId: queryHistoryUuid,
-                    projectId: projectUuid,
-                    warehouseType: warehouseClient.credentials.type,
-                    warehouseExecutionTimeMs: durationMs,
-                    columnsCount: Object.keys(fieldsMap).length,
-                    totalRowCount: totalRows,
-                },
-            });
-
-            // Wait for the cache to be written before marking the query as ready
-            if (resultsCache) {
-                await resultsCache.close();
-                await this.cacheService?.updateCache(
-                    resultsCache.cacheKey,
-                    projectUuid,
-                    {
-                        status: ResultsCacheStatus.READY,
-                        total_row_count: totalRows,
-                    },
-                );
-                this.analytics.track({
-                    userId: user.userUuid,
-                    event: 'results_cache.write',
-                    properties: {
-                        queryId: queryHistoryUuid,
-                        projectId: projectUuid,
-                        cacheKey: resultsCache.cacheKey,
-                        totalRowCount: totalRows,
-                    },
-                });
-            }
-
-            await this.queryHistoryModel.update(
-                queryHistoryUuid,
-                projectUuid,
-                user.userUuid,
-                {
-                    warehouse_query_id: queryId,
-                    warehouse_query_metadata: queryMetadata,
-                    status: QueryHistoryStatus.READY,
-                    error: null,
-                    warehouse_execution_time_ms:
-                        durationMs !== null ? Math.round(durationMs) : null,
-                    total_row_count: totalRows,
-                },
-            );
-        } catch (e) {
-            this.analytics.track({
-                userId: user.userUuid,
-                event: 'query.error',
-                properties: {
-                    queryId: queryHistoryUuid,
-                    projectId: projectUuid,
-                    warehouseType: warehouseClient.credentials.type,
-                },
-            });
-            await this.queryHistoryModel.update(
-                queryHistoryUuid,
-                projectUuid,
-                user.userUuid,
-                {
-                    status: QueryHistoryStatus.ERROR,
-                    error: getErrorMessage(e),
-                },
-            );
-
-            // When the query fails, we delete the cache entry
-            if (resultsCache) {
-                await this.cacheService?.deleteCache(
-                    resultsCache.cacheKey,
-                    projectUuid,
-                );
-                this.analytics.track({
-                    userId: user.userUuid,
-                    event: 'results_cache.delete',
-                    properties: {
-                        queryId: queryHistoryUuid,
-                        projectId: projectUuid,
-                        cacheKey: resultsCache.cacheKey,
-                    },
-                });
-            }
-        } finally {
-            void sshTunnel.disconnect();
-
-            if (resultsCache) {
-                void resultsCache.close();
-            }
-        }
-    }
-
-    async executeAsyncQuery(
-        args: ExecuteAsyncMetricQueryArgs & {
-            queryTags: RunQueryTags;
-            explore: Explore;
-        },
-        requestParameters: ExecuteAsyncQueryRequestParams,
-        pivotConfiguration?: {
-            indexColumn: PivotIndexColum;
-            valuesColumns: ValuesColumn[];
-            groupByColumns: GroupByColumn[] | undefined;
-            sortBy: SortBy | undefined;
-        },
-    ): Promise<ExecuteAsyncQueryReturn> {
-        return wrapSentryTransaction(
-            'ProjectService.executeAsyncQuery',
-            {},
-            async (span) => {
-                const {
-                    user,
-                    projectUuid,
-                    context,
-                    granularity,
-                    queryTags,
-                    explore,
-                } = args;
-
-                try {
-                    if (!isUserWithOrg(user)) {
-                        throw new ForbiddenError(
-                            'User is not part of an organization',
-                        );
-                    }
-
-                    const { organizationUuid } =
-                        await this.projectModel.getSummary(projectUuid);
-
-                    if (
-                        user.ability.cannot(
-                            'view',
-                            subject('Project', {
-                                organizationUuid,
-                                projectUuid,
-                            }),
-                        )
-                    ) {
-                        throw new ForbiddenError();
-                    }
-
-                    const { warehouseClient, sshTunnel } =
-                        await this._getWarehouseClient(
-                            projectUuid,
-                            await this.getWarehouseCredentials(
-                                projectUuid,
-                                user.userUuid,
-                            ),
-                            {
-                                snowflakeVirtualWarehouse: explore.warehouse,
-                                databricksCompute: explore.databricksCompute,
-                            },
-                        );
-
-                    span.setAttribute('lightdash.projectUuid', projectUuid);
-                    span.setAttribute(
-                        'warehouse.type',
-                        warehouseClient.credentials.type,
-                    );
-
-                    const { metricQuery } = args;
-                    const userAttributes =
-                        await this.userAttributesModel.getAttributeValuesForOrgMember(
-                            {
-                                organizationUuid,
-                                userUuid: user.userUuid,
-                            },
-                        );
-
-                    const emailStatus =
-                        await this.emailModel.getPrimaryEmailStatus(
-                            user.userUuid,
-                        );
-                    const intrinsicUserAttributes = emailStatus.isVerified
-                        ? getIntrinsicUserAttributes(user)
-                        : {};
-
-                    const fullQuery = await ProjectService._compileQuery(
-                        metricQuery,
-                        explore,
-                        warehouseClient,
-                        intrinsicUserAttributes,
-                        userAttributes,
-                        this.lightdashConfig.query.timezone || 'UTC',
-                        granularity,
-                    );
-
-                    const {
-                        query: compiledQuery,
-                        fields: fieldsFromQuery,
-                        hasExampleMetric,
-                    } = fullQuery;
-
-                    let pivotedQuery = null;
-                    if (pivotConfiguration) {
-                        pivotedQuery =
-                            await ProjectService.applyPivotToSqlQuery({
-                                warehouseType: warehouseClient.credentials.type,
-                                sql: compiledQuery,
-                                indexColumn: pivotConfiguration.indexColumn,
-                                valuesColumns: pivotConfiguration.valuesColumns,
-                                groupByColumns:
-                                    pivotConfiguration.groupByColumns,
-                                sortBy: pivotConfiguration.sortBy,
-                            });
-                    }
-
-                    const query = pivotedQuery || compiledQuery;
-
-                    const fieldsWithOverrides: ItemsMap = Object.fromEntries(
-                        Object.entries(fieldsFromQuery).map(([key, value]) => {
-                            if (
-                                metricQuery.metricOverrides &&
-                                metricQuery.metricOverrides[key]
-                            ) {
-                                return [
-                                    key,
-                                    {
-                                        ...value,
-                                        ...metricQuery.metricOverrides[key],
-                                    },
-                                ];
-                            }
-                            return [key, value];
-                        }),
-                    );
-
-                    const fieldsMap = fieldsWithOverrides;
-                    span.setAttribute('generatedSql', query);
-
-                    const onboardingRecord =
-                        await this.onboardingModel.getByOrganizationUuid(
-                            user.organizationUuid,
-                        );
-
-                    if (!onboardingRecord.ranQueryAt) {
-                        await this.onboardingModel.update(
-                            user.organizationUuid,
-                            {
-                                ranQueryAt: new Date(),
-                            },
-                        );
-                    }
-
-                    const resultsCacheEnabled =
-                        this.lightdashConfig.resultsCache?.resultsEnabled;
-
-                    let resultsCache: CreateCacheResult | undefined;
-
-                    if (resultsCacheEnabled && this.cacheService) {
-                        resultsCache =
-                            await this.cacheService.createOrGetExistingCache(
-                                projectUuid,
-                                {
-                                    sql: query,
-                                    timezone: metricQuery.timezone,
-                                },
-                                args.invalidateCache,
-                            );
-
-                        if (!resultsCache.cacheHit) {
-                            this.analytics.track({
-                                userId: user.userUuid,
-                                event: 'results_cache.create',
-                                properties: {
-                                    projectId: projectUuid,
-                                    cacheKey: resultsCache.cacheKey,
-                                    totalRowCount: resultsCache.totalRowCount,
-                                    createdAt: resultsCache.createdAt,
-                                    expiresAt: resultsCache.expiresAt,
-                                },
-                            });
-                        }
-                    }
-
-                    const { queryUuid: queryHistoryUuid } =
-                        await this.queryHistoryModel.create({
-                            projectUuid,
-                            organizationUuid,
-                            createdByUserUuid: user.userUuid,
-                            context,
-                            fields: fieldsMap,
-                            compiledSql: query,
-                            requestParameters,
-                            metricQuery,
-                            cacheKey: resultsCache?.cacheKey || null,
-                        });
-
-                    this.analytics.track({
-                        userId: user.userUuid,
-                        event: 'query.executed',
-                        properties: {
-                            organizationId: organizationUuid,
-                            projectId: projectUuid,
-                            context,
-                            queryId: queryHistoryUuid,
-                            warehouseType: warehouseClient.credentials.type,
-                            ...ProjectService.getMetricQueryExecutionProperties(
-                                {
-                                    metricQuery,
-                                    hasExampleMetric,
-                                    queryTags,
-                                    granularity,
-                                    chartUuid:
-                                        'chartUuid' in requestParameters
-                                            ? requestParameters.chartUuid
-                                            : undefined,
-                                    explore,
-                                },
-                            ),
-                            cacheMetadata: {
-                                cacheHit: resultsCache?.cacheHit || false,
-                                cacheUpdatedTime: resultsCache?.updatedAt,
-                                cacheExpiresAt: resultsCache?.expiresAt,
-                            },
-                        },
-                    });
-
-                    if (resultsCache?.cacheHit) {
-                        await this.queryHistoryModel.update(
-                            queryHistoryUuid,
-                            projectUuid,
-                            user.userUuid,
-                            {
-                                // If the cache is ready, we set the query history status to READY
-                                // Otherwise, we set it to PENDING
-                                status:
-                                    resultsCache.status ===
-                                    ResultsCacheStatus.READY
-                                        ? QueryHistoryStatus.READY
-                                        : QueryHistoryStatus.PENDING,
-                                error: null,
-                                total_row_count: resultsCache.totalRowCount,
-                                warehouse_execution_time_ms: 0, // When cache is hit, no query is executed
-                            },
-                        );
-
-                        return {
-                            queryUuid: queryHistoryUuid,
-                            cacheMetadata: {
-                                cacheHit: resultsCache.cacheHit,
-                                cacheUpdatedTime: resultsCache.updatedAt,
-                                cacheExpiresAt: resultsCache.expiresAt,
-                            },
-                        } satisfies ExecuteAsyncQueryReturn;
-                    }
-
-                    // Trigger query in the background, update query history and cache when complete
-                    void this.runAsyncWarehouseQuery({
-                        user,
-                        projectUuid,
-                        query,
-                        fieldsMap,
-                        queryTags,
-                        warehouseClient,
-                        sshTunnel,
-                        queryHistoryUuid,
-                        // resultsCache is either MissCacheResult or undefined at this point,
-                        // meaning that the cache was not hit or that cache is not enabled
-                        resultsCache,
-                    });
-
-                    return {
-                        queryUuid: queryHistoryUuid,
-                        cacheMetadata: {
-                            cacheHit: resultsCache?.cacheHit || false,
-                            cacheUpdatedTime: resultsCache?.updatedAt,
-                            cacheExpiresAt: resultsCache?.expiresAt,
-                        },
-                    } satisfies ExecuteAsyncQueryReturn;
-                } catch (e) {
-                    span.setStatus({
-                        code: 2, // ERROR
-                        message: getErrorMessage(e),
-                    });
-                    throw e;
-                } finally {
-                    span.end();
-                }
-            },
-        );
-    }
-
-    // execute
-    async executeAsyncMetricQuery({
-        user,
-        projectUuid,
-        granularity,
-        context,
-        metricQuery,
-        invalidateCache,
-    }: ExecuteAsyncMetricQueryArgs): Promise<ApiExecuteAsyncQueryResults> {
-        if (!isUserWithOrg(user)) {
-            throw new ForbiddenError('User is not part of an organization');
-        }
-        const { organizationUuid } =
-            await this.projectModel.getWithSensitiveFields(projectUuid);
-
-        if (
-            user.ability.cannot(
-                'manage',
-                subject('Explore', { organizationUuid, projectUuid }),
-            )
-        ) {
-            throw new ForbiddenError();
-        }
-
-        if (
-            metricQuery.customDimensions?.some(isCustomSqlDimension) &&
-            user.ability.cannot(
-                'manage',
-                subject('CustomSql', { organizationUuid, projectUuid }),
-            )
-        ) {
-            throw new ForbiddenError(
-                'User cannot run queries with custom SQL dimensions',
-            );
-        }
-
-        const requestParameters: ExecuteAsyncMetricQueryRequestParams = {
-            context,
-            query: metricQuery,
-        };
-
-        const queryTags: RunQueryTags = {
-            organization_uuid: organizationUuid,
-            project_uuid: projectUuid,
-            user_uuid: user.userUuid,
-            explore_name: metricQuery.exploreName,
-            query_context: context,
-        };
-
-        const explore = await this.getExplore(
-            user,
-            projectUuid,
-            metricQuery.exploreName,
-            organizationUuid,
-        );
-
-        const { queryUuid, cacheMetadata } = await this.executeAsyncQuery(
-            {
-                user,
-                metricQuery,
-                projectUuid,
-                explore,
-                context,
-                queryTags,
-                granularity,
-                invalidateCache,
-            },
-            requestParameters,
-        );
-
-        return {
-            queryUuid,
-            cacheMetadata,
-            appliedDashboardFilters: null,
-        };
-    }
-
-    async executeAsyncSqlQuery({
-        user,
-        projectUuid,
-        sql,
-        context,
-        invalidateCache,
-        pivotConfiguration,
-    }: ExecuteAsyncSqlQueryArgs): Promise<ApiExecuteAsyncQueryResults> {
-        if (!isUserWithOrg(user)) {
-            throw new ForbiddenError('User does not belong to an organization');
-        }
-
-        const { organizationUuid } = await this.projectModel.getSummary(
-            projectUuid,
-        );
-
-        if (
-            user.ability.cannot(
-                'manage',
-                subject('SqlRunner', {
-                    organizationUuid,
-                    projectUuid,
-                }),
-            )
-        ) {
-            throw new ForbiddenError();
-        }
-
-        const { warehouseClient } = await this._getWarehouseClient(
-            projectUuid,
-            await this.getWarehouseCredentials(projectUuid, user.userUuid),
-        );
-
-        const queryTags: RunQueryTags = {
-            organization_uuid: organizationUuid,
-            user_uuid: user.userUuid,
-            query_context: context,
-        };
-
-        // Get one row to get the column definitions
-        const columns: { name: string; type: DimensionType }[] = [];
-        await warehouseClient.streamQuery(
-            applyLimitToSqlQuery({ sqlQuery: sql, limit: 1 }),
-            (row) => {
-                if (row.fields) {
-                    Object.keys(row.fields).forEach((key) => {
-                        columns.push({
-                            name: key,
-                            type: row.fields[key].type,
-                        });
-                    });
-                }
-            },
-            {
-                tags: queryTags,
-            },
-        );
-
-        const vizColumns = columns.map((col) => ({
-            reference: col.name,
-            type: col.type,
-        }));
-
-        const virtualView = createVirtualViewObject(
-            'virtual_view',
-            sql,
-            vizColumns,
-            warehouseClient,
-        );
-
-        const dimensions = Object.values(
-            virtualView.tables[virtualView.baseTable].dimensions,
-        ).map((d) => convertFieldRefToFieldId(d.name, virtualView.name));
-
-        const prefixedPivotConfiguration = pivotConfiguration
-            ? prefixPivotConfigurationReferences(
-                  pivotConfiguration,
-                  `${virtualView.name}`,
-              )
-            : undefined;
-
-        const query: MetricQuery = {
-            exploreName: virtualView.name,
-            dimensions,
-            metrics: [],
-            filters: {},
-            tableCalculations: [],
-            sorts: [],
-            customDimensions: [],
-            additionalMetrics: [],
-            limit: 500,
-        };
-
-        const { queryUuid, cacheMetadata } = await this.executeAsyncQuery(
-            {
-                user,
-                projectUuid,
-                explore: virtualView,
-                queryTags,
-                metricQuery: query,
-                context,
-            },
-            {
-                query,
-                invalidateCache,
-            },
-            prefixedPivotConfiguration,
-        );
-
-        return {
-            queryUuid,
-            cacheMetadata,
-            appliedDashboardFilters: null,
-        };
-    }
-
-    async executeAsyncSavedChartQuery({
-        user,
-        projectUuid,
-        chartUuid,
-        versionUuid,
-        context,
-        invalidateCache,
-    }: ExecuteAsyncSavedChartQueryArgs): Promise<ApiExecuteAsyncQueryResults> {
-        // Check user is in organization
-        if (!isUserWithOrg(user)) {
-            throw new ForbiddenError('User does not belong to an organization');
-        }
-
-        const {
-            uuid: savedChartUuid,
-            organizationUuid: savedChartOrganizationUuid,
-            projectUuid: savedChartProjectUuid,
-            spaceUuid: savedChartSpaceUuid,
-            tableName: savedChartTableName,
-            metricQuery,
-        } = await this.savedChartModel.get(chartUuid, versionUuid);
-
-        // Check chart belongs to project
-        if (savedChartProjectUuid !== projectUuid) {
-            throw new ForbiddenError('Chart does not belong to project');
-        }
-
-        const space = await this.spaceModel.getSpaceSummary(
-            savedChartSpaceUuid,
-        );
-
-        const access = await this.spaceModel.getUserSpaceAccess(
-            user.userUuid,
-            space.uuid,
-        );
-
-        if (
-            user.ability.cannot(
-                'view',
-                subject('SavedChart', {
-                    organizationUuid: savedChartOrganizationUuid,
-                    projectUuid,
-                    isPrivate: space.isPrivate,
-                    access,
-                }),
-            ) ||
-            user.ability.cannot(
-                'view',
-                subject('Project', {
-                    organizationUuid: savedChartOrganizationUuid,
-                    projectUuid,
-                }),
-            )
-        ) {
-            throw new ForbiddenError();
-        }
-
-        await this.analyticsModel.addChartViewEvent(
-            savedChartUuid,
-            user.userUuid,
-        );
-
-        const requestParameters: ExecuteAsyncSavedChartRequestParams = {
-            context,
-            chartUuid,
-            versionUuid,
-        };
-
-        const queryTags: RunQueryTags = {
-            organization_uuid: savedChartOrganizationUuid,
-            project_uuid: projectUuid,
-            user_uuid: user.userUuid,
-            chart_uuid: chartUuid,
-            explore_name: savedChartTableName,
-            query_context: context,
-        };
-
-        const explore = await this.getExplore(
-            user,
-            projectUuid,
-            savedChartTableName,
-            savedChartOrganizationUuid,
-        );
-
-        const { queryUuid, cacheMetadata } = await this.executeAsyncQuery(
-            {
-                user,
-                projectUuid,
-                explore,
-                context,
-                queryTags,
-                invalidateCache,
-                metricQuery,
-            },
-            requestParameters,
-        );
-
-        return {
-            queryUuid,
-            cacheMetadata,
-            appliedDashboardFilters: null,
-        };
-    }
-
-    async executeAsyncDashboardChartQuery({
-        user,
-        projectUuid,
-        chartUuid,
-        dashboardUuid,
-        dashboardFilters,
-        dashboardSorts,
-        granularity,
-        context,
-        invalidateCache,
-    }: ExecuteAsyncDashboardChartQueryArgs): Promise<ApiExecuteAsyncQueryResults> {
-        if (!isUserWithOrg(user)) {
-            throw new ForbiddenError('User is not part of an organization');
-        }
-
-        const savedChart = await this.savedChartModel.get(chartUuid);
-        const { organizationUuid, projectUuid: savedChartProjectUuid } =
-            savedChart;
-
-        if (savedChartProjectUuid !== projectUuid) {
-            throw new ForbiddenError('Dashboard does not belong to project');
-        }
-
-        const [space, explore] = await Promise.all([
-            this.spaceModel.getSpaceSummary(savedChart.spaceUuid),
-            this.getExplore(
-                user,
-                projectUuid,
-                savedChart.tableName,
-                organizationUuid,
-            ),
-        ]);
-
-        const access = await this.spaceModel.getUserSpaceAccess(
-            user.userUuid,
-            space.uuid,
-        );
-
-        if (
-            user.ability.cannot(
-                'view',
-                subject('SavedChart', {
-                    organizationUuid,
-                    projectUuid,
-                    isPrivate: space.isPrivate,
-                    access,
-                }),
-            ) ||
-            user.ability.cannot(
-                'view',
-                subject('Project', {
-                    organizationUuid,
-                    projectUuid,
-                }),
-            )
-        ) {
-            throw new ForbiddenError();
-        }
-
-        await this.analyticsModel.addChartViewEvent(
-            savedChart.uuid,
-            user.userUuid,
-        );
-
-        const tables = Object.keys(explore.tables);
-        const appliedDashboardFilters: DashboardFilters = {
-            dimensions: getDashboardFilterRulesForTables(
-                tables,
-                dashboardFilters.dimensions,
-            ),
-            metrics: getDashboardFilterRulesForTables(
-                tables,
-                dashboardFilters.metrics,
-            ),
-            tableCalculations: getDashboardFilterRulesForTables(
-                tables,
-                dashboardFilters.tableCalculations,
-            ),
-        };
-
-        const metricQueryWithDashboardOverrides: MetricQuery = {
-            ...addDashboardFiltersToMetricQuery(
-                savedChart.metricQuery,
-                appliedDashboardFilters,
-                explore,
-            ),
-            sorts:
-                dashboardSorts && dashboardSorts.length > 0
-                    ? dashboardSorts
-                    : savedChart.metricQuery.sorts,
-        };
-
-        const exploreDimensions = getDimensions(explore);
-
-        const metricQueryDimensions = [
-            ...metricQueryWithDashboardOverrides.dimensions,
-            ...(metricQueryWithDashboardOverrides.customDimensions ?? []),
-        ];
-        const hasADateDimension = exploreDimensions.find(
-            (c) =>
-                metricQueryDimensions.includes(getItemId(c)) && isDateItem(c),
-        );
-
-        if (hasADateDimension) {
-            metricQueryWithDashboardOverrides.metadata = {
-                hasADateDimension: {
-                    name: hasADateDimension.name,
-                    label: hasADateDimension.label,
-                },
-            };
-        }
-
-        const requestParameters: ExecuteAsyncDashboardChartRequestParams = {
-            context,
-            chartUuid,
-            dashboardUuid,
-            dashboardFilters,
-            dashboardSorts,
-            granularity,
-        };
-
-        const queryTags: RunQueryTags = {
-            organization_uuid: organizationUuid,
-            project_uuid: projectUuid,
-            user_uuid: user.userUuid,
-            chart_uuid: chartUuid,
-            dashboard_uuid: dashboardUuid,
-            explore_name: explore.name,
-            query_context: context,
-        };
-
-        const { queryUuid, cacheMetadata } = await this.executeAsyncQuery(
-            {
-                user,
-                projectUuid,
-                explore,
-                metricQuery: metricQueryWithDashboardOverrides,
-                context,
-                queryTags,
-                invalidateCache,
-                granularity,
-            },
-            requestParameters,
-        );
-
-        return {
-            queryUuid,
-            cacheMetadata,
-            appliedDashboardFilters,
-        };
-    }
-
-    async executeAsyncUnderlyingDataQuery({
-        user,
-        projectUuid,
-        underlyingDataSourceQueryUuid,
-        filters,
-        underlyingDataItemId,
-        context,
-        invalidateCache,
-    }: ExecuteAsyncUnderlyingDataQueryArgs): Promise<ApiExecuteAsyncQueryResults> {
-        if (!isUserWithOrg(user)) {
-            throw new ForbiddenError('User is not part of an organization');
-        }
-        const { organizationUuid } =
-            await this.projectModel.getWithSensitiveFields(projectUuid);
-
-        if (
-            user.ability.cannot(
-                'view',
-                subject('UnderlyingData', { organizationUuid, projectUuid }),
-            )
-        ) {
-            throw new ForbiddenError();
-        }
-
-        const { metricQuery, fields: metricQueryFields } =
-            await this.queryHistoryModel.get(
-                underlyingDataSourceQueryUuid,
-                projectUuid,
-                user.userUuid,
-            );
-
-        const { exploreName } = metricQuery;
-
-        const explore = await this.getExplore(
-            user,
-            projectUuid,
-            exploreName,
-            organizationUuid,
-        );
-
-        const underlyingDataItem = underlyingDataItemId
-            ? metricQueryFields[underlyingDataItemId]
-            : undefined;
-
-        const joinedTables = explore.joinedTables.map(
-            (joinedTable) => joinedTable.table,
-        );
-
-        const availableTables = new Set([
-            ...joinedTables,
-            ...Object.values(metricQueryFields)
-                .filter(isField)
-                .map((field) => field.table),
-        ]);
-
-        const itemShowUnderlyingValues =
-            isField(underlyingDataItem) && isMetric(underlyingDataItem)
-                ? underlyingDataItem.showUnderlyingValues
-                : undefined;
-
-        const itemShowUnderlyingTable = isField(underlyingDataItem)
-            ? underlyingDataItem.table
-            : undefined;
-
-        const allDimensions = [
-            ...(metricQuery.customDimensions?.filter(
-                (dimension) => !isCustomBinDimension(dimension),
-            ) || []),
-            ...getDimensions(explore),
-        ];
-
-        const isValidNonCustomDimension = (
-            dimension: CustomDimension | CompiledDimension,
-        ) =>
-            !isCustomDimension(dimension) &&
-            !dimension.timeInterval &&
-            !dimension.hidden;
-
-        const availableDimensions = allDimensions.filter(
-            (dimension) =>
-                availableTables.has(dimension.table) &&
-                (isValidNonCustomDimension(dimension) ||
-                    isCustomDimension(dimension)) &&
-                (itemShowUnderlyingValues !== undefined
-                    ? (itemShowUnderlyingValues.includes(dimension.name) &&
-                          itemShowUnderlyingTable === dimension.table) ||
-                      itemShowUnderlyingValues.includes(
-                          `${dimension.table}.${dimension.name}`,
-                      )
-                    : true),
-        );
-
-        const requestParameters: ExecuteAsyncUnderlyingDataRequestParams = {
-            context,
-            underlyingDataSourceQueryUuid,
-            filters,
-            underlyingDataItemId,
-        };
-
-        const queryTags: RunQueryTags = {
-            organization_uuid: organizationUuid,
-            project_uuid: projectUuid,
-            user_uuid: user.userUuid,
-            explore_name: exploreName,
-            query_context: context,
-        };
-
-        const underlyingDataMetricQuery: MetricQuery = {
-            exploreName,
-            dimensions: availableDimensions.map(getItemId),
-            // Remove custom bin dimensions from underlying data query
-            customDimensions: metricQuery.customDimensions?.filter(
-                (dimension) => !isCustomBinDimension(dimension),
-            ),
-            filters,
-            metrics: [],
-            sorts: [],
-            limit: 500,
-            tableCalculations: [],
-            additionalMetrics: [],
-        };
-
-        const { queryUuid: underlyingDataQueryUuid, cacheMetadata } =
-            await this.executeAsyncQuery(
-                {
-                    user,
-                    metricQuery: underlyingDataMetricQuery,
-                    projectUuid,
-                    explore,
-                    context,
-                    queryTags,
-                    invalidateCache,
-                },
-                requestParameters,
-            );
-
-        return {
-            queryUuid: underlyingDataQueryUuid,
-            appliedDashboardFilters: null,
-            cacheMetadata,
-        };
     }
 
     private async runQueryAndFormatRows({
